+++ NEWS.md
@@ -1,9 +1,10 @@
-climatrends 0.1.4 (2020-05-07)
+climatrends 0.1.5 (2020-05-14)
 =========================
 
 ### NEW FEATURES
 
-* New function `anthesis()` to compute crop sensitive indices.
+* New function `crop_sensitive()` to compute crop sensitive indices.
+* New function `late_frost()` to compute late spring frost.
 * Export function `get_timeseries()`
 * Enable future modular expansion of data sources using API clients by creating specific functions that are called with the argument `data.from` in `get_timeseries()`
 * Enable argument `last.day`, optional to `span`, when indices are retrieved with focus on time series analysis